Add github actions
This adds three main workflows for github actions:
 * Typical rust build/test/lint
 * Cross-build to AArch64
 * A security audit job

The linters used are the nightly variants of clippy, and rustfmt
as they have a much greater array of features compared to their
stable counterparts.

All compilations are performed under the latest stable toolchain.

The security auditing workflow is excuted every midnight and
checks all dependencies of the project for any reported security
vulnerabilties on the RustSec advisories database. If found it
will create a github issue.

These workflows are all taken from the actions-rs repository
and fall under the MIT license.